v1.0 : чуть редачен код
